# apps/api-gateway/doc/openapi.yml
...
openapi: 3.0.3
info:
  title: Movie Service API
  version: 1.0.0
  description: Quản lý thông tin website đặt vé xem phim
servers:
  - url: /api/{version}
    variables:
      version:
        default: v1
        enum:
          - v1
          - v2
...
paths:
...
  /genres:
    get:
      tags: [Genres]
      summary: Lấy danh sách thể loại phim
      responses:
        '200':
          description: Danh sách thể loại phim
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiSuccessResponse'
                  - type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/GenreResponse'
        '400':
          description: Lỗi request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
    post:
      tags: [Genres]
      summary: Cập nhật thể loại phim
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenreResponse'
      responses:
        '200':
          description: Thể loại phim được cập nhật thành công
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiSuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/GenreResponse'
        '400':
          description: Lỗi request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
  /genres/{genreId}:
    put:
      tags: [Genres]
      summary: Cập nhật thể loại phim
      parameters:
        - name: genreId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenreResponse'
      responses:
        '200':
          description: Thể loại phim được cập nhật thành công
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiSuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/GenreResponse'
        '400':
          description: Lỗi request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
    delete:
      tags: [Genres]
      summary: Xóa thể loại phim
      parameters:
        - name: genreId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Thể loại phim được xóa thành công
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiSuccessResponse'
        '400':
          description: Lỗi request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'

  ## Cinema Endpoints
  # ----------------------
...
  /showtimes/{id}/seats:
    get:
      summary: Get showtime seats
      tags:
        - Showtimes
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Showtime seats retrieved
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiSuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/ShowtimeSeatResponse'
        '404':
          description: Showtime not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
  /showtimes/showtime/{showtimeId}/ttl:
    get:
      summary: Get TTL of user's session for a showtime
      tags:
        - Showtimes
      parameters:
        - name: showtimeId
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: TTL retrieved
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiSuccessResponse'
                  - type: object
                    properties:
                      ttl:
                        type: integer
                        example: 100
        '404':
          description: Showtime not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
  /showtimes/showtime:
    post:
      summary: Create a new showtime
      tags:
        - Showtimes
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateShowtimeRequest'
      responses:
        '200':
          description: Showtime created
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiSuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/ShowtimeDetailResponse'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
  /showtimes/batch:
    post:
      summary: Batch create showtimes
      tags:
        - Showtimes
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BatchCreateShowtimesInput'
      responses:
        '200':
          description: Batch created
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiSuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/BatchCreateShowtimeResponse'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
  /showtimes/showtime/{id}:
    patch:
      summary: Update showtime
      tags:
        - Showtimes
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateShowtimeRequest'
      responses:
        '200':
          description: Showtime updated
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiSuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/ShowtimeDetailResponse'
        '404':
          description: Showtime not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'

    delete:
      summary: Delete showtime
      tags:
        - Showtimes
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Showtime deleted
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiSuccessResponse'
                  - type: object
                    properties:
                      message:
                        type: string
                        example: Showtime deleted successfully
        '404':
          description: Showtime not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
  /ticket-pricings/hall/{hallId}:
    get:
      summary: Get ticket pricing list for a hall
      tags:
        - Ticket Pricing
      parameters:
        - name: hallId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Ticket pricing retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiSuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/TicketPricingResponse'
        '404':
          description: Hall not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
  /ticket-pricings/pricing/{pricingId}:
    patch:
      summary: Update ticket pricing
      tags:
        - Ticket Pricing
      parameters:
        - name: pricingId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                price:
                  type: number
                  example: 120000
              required:
                - price
      responses:
        '200':
          description: Ticket pricing updated successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiSuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/TicketPricingResponse'
        '404':
          description: Pricing not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'

  # ----------------------
  # PROMOTIONS endpoints
  # ----------------------
...
components:
  schemas:
...
    ErrorItem:
      type: object
      properties:
        code: { type: string }
        message: { type: string }
    ApiErrorResponse:
      type: object
      required: [success, message, errors, timestamp, path, statusCode]
      properties:
        success: { type: boolean, enum: [false] }
        message: { type: string }
        errors:
          type: array
          nullable: true
          items: { $ref: '#/components/schemas/ErrorItem' }
        timestamp: { type: string, format: date-time }
        path: { type: string }
        statusCode: { type: integer }
    ApiSuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          nullable: true
        data:
          nullable: true
        meta:
          type: object
          nullable: true
        timestamp:
          type: string
          format: date-time
        path: { type: string }
      required: [success, message, timestamp]
...
    GenreResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
...
    ShowtimeSeatResponse:
      type: object
      properties:
        showtime:
          type: object
          properties:
            id: { type: string }
            movieId: { type: string }
            movieTitle: { type: string }
            start_time: { type: string, format: date-time }
            end_time: { type: string, format: date-time }
            dateType:
              type: string
              enum: [WEEKDAY, WEEKEND, HOLIDAY]
            format:
              type: string
              enum: [TWO_D, THREE_D, IMAX, FOUR_DX]
            language: { type: string }
            subtitles:
              type: array
              items: { type: string }
        cinemaId: { type: string }
        cinemaName: { type: string }
        hallId: { type: string }
        hallName: { type: string }
        layoutType:
          type: string
          enum: [STANDARD, DUAL_AISLE, STADIUM]
        seat_map:
          type: array
          items:
            type: object
            properties:
              row: { type: string }
              seats:
                type: array
                items:
                  type: object
                  properties:
                    id: { type: string }
                    number: { type: number }
                    seatType:
                      type: string
                      enum: [STANDARD, VIP, COUPLE, PREMIUM, WHEELCHAIR]
                    seatStatus:
                      type: string
                      enum: [ACTIVE, BROKEN, MAINTENANCE]
                    reservationStatus:
                      type: string
                      enum: [AVAILABLE, HELD, CONFIRMED, CANCELLED]
                    isHeldByCurrentUser:
                      type: boolean
                      nullable: true
        ticketPrices:
          type: array
          items:
            type: object
            properties:
              seatType:
                type: string
                enum: [STANDARD, VIP, COUPLE, PREMIUM, WHEELCHAIR]
              price: { type: number }
        rules:
          type: object
          properties:
            max_selectable: { type: number }
            hold_time_seconds: { type: number }
    CreateShowtimeRequest:
      type: object
      properties:
        movieId: { type: string }
        movieReleaseId: { type: string }
        cinemaId: { type: string }
        hallId: { type: string }
        startTime:
          type: string
          format: date-time
        format:
          type: string
          enum: [TWO_D, THREE_D, IMAX, FOUR_DX]
        language: { type: string }
        subtitles:
          type: array
          items: { type: string }
    UpdateShowtimeRequest:
      type: object
      properties:
        movieId: { type: string }
        movieReleaseId: { type: string }
        cinemaId: { type: string }
        hallId: { type: string }
        startTime:
          type: string
          format: date-time
        format:
          type: string
          enum: [TWO_D, THREE_D, IMAX, FOUR_DX]
        language: { type: string }
        subtitles:
          type: array
          items: { type: string }
    ShowtimeDetailResponse:
      type: object
      properties:
        id: { type: string }
        movieId: { type: string }
        movieReleaseId: { type: string, nullable: true }
        cinemaId: { type: string }
        hallId: { type: string }
        startTime: { type: string, format: date-time }
        endTime: { type: string, format: date-time }
        format:
          type: string
          enum: [TWO_D, THREE_D, IMAX, FOUR_DX]
        language: { type: string }
        subtitles:
          type: array
          items: { type: string }
        availableSeats: { type: number }
        totalSeats: { type: number }
        status:
          type: string
          enum: [SELLING, STOPPED, CANCELLED]
        dayType:
          type: string
          enum: [WEEKDAY, WEEKEND, HOLIDAY]
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    BatchCreateShowtimesInput:
      type: object
      properties:
        movieId: { type: string }
        movieReleaseId: { type: string }
        cinemaId: { type: string }
        hallId: { type: string }
        startDate: { type: string, format: date-time }
        endDate: { type: string, format: date-time }
        timeSlots:
          type: array
          items:
            type: string
            pattern: '^([01]\d|2[0-3]):([0-5]\d)$'
          example: ['14:00', '18:00', '21:00']
        repeatType:
          type: string
          enum: [DAILY, WEEKLY, CUSTOM_WEEKDAYS]
        weekdays:
          type: array
          description: Thứ trong tuần
          items:
            type: integer
            enum: [0, 1, 2, 3, 4, 5, 6]
          example: [1, 3, 5]
        format:
          type: string
          enum: [TWO_D, THREE_D, IMAX, FOUR_DX]
        language: { type: string }
        subtitles:
          type: array
          items: { type: string }
    BatchCreateShowtimeResponse:
      type: object
      properties:
        createdCount: { type: number }
        skippedCount: { type: number }
        created:
          type: array
          items:
            $ref: '#/components/schemas/ShowtimeDetailResponse'
        skipped:
          type: array
          items:
            type: object
            properties:
              start: { type: string, format: date-time }
              reason: { type: string, example: 'TIME_CONFLICT' }

    # ----------------------
    # TICKET PRICING SCHEMAS
    # ----------------------
    TicketPricingResponse:
      type: object
      properties:
        id:
          type: string
        hallId:
          type: string
        seatType:
          type: string
          enum: [STANDARD, VIP, COUPLE, PREMIUM, WHEELCHAIR]
        dayType:
          type: string
          enum: [WEEKDAY, WEEKEND, HOLIDAY]
        price:
          type: number

    # ----------------------
    # PROMOTIONS SCHEMAS
    # ----------------------
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT